ReOrganized classes and added launcher icon.
