--- app/src/main/java/com/papfree/bloodlife/GetUserCallBack.java
@@ -1,10 +1,0 @@
-package com.papfree.bloodlife;
-
-/**
- * Created by Aravind on 03/06/2016.
- */
-interface GetUserCallBack {
-
-    public abstract void done(User returnedUser);
-
-}
