+++ app/src/main/java/com/papfree/bloodlife/Model/User.java
@@ -1,4 +1,4 @@
-package com.papfree.bloodlife;
+package com.papfree.bloodlife.Model;
 
 public class User {
