--- app/src/main/java/com/papfree/bloodlife/ServerRequests.java
@@ -1,148 +1,0 @@
-package com.papfree.bloodlife;
-
-import android.app.ProgressDialog;
-import android.content.Context;
-import android.os.AsyncTask;
-
-import org.apache.http.HttpEntity;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.entity.UrlEncodedFormEntity;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.impl.client.DefaultHttpClient;
-import org.apache.http.message.BasicNameValuePair;
-import org.apache.http.params.BasicHttpParams;
-import org.apache.http.params.HttpConnectionParams;
-import org.apache.http.params.HttpParams;
-import org.apache.http.util.EntityUtils;
-import org.json.JSONObject;
-
-import java.util.ArrayList;
-
-public class ServerRequests {
-
-    ProgressDialog progressDialog;
-    public static final int CONNECTION_TIMEOUT = 1000 * 15;
-    public static final String SERVER_ADDRESS = "http://papfree.site88.net/";
-
-    public ServerRequests(Context context) {
-        progressDialog = new ProgressDialog(context);
-        progressDialog.setCancelable(false);
-        progressDialog.setTitle("Processing");
-        progressDialog.setMessage("Please Wait...");
-    }
-
-    public void storeUserDataInBackground(User user, GetUserCallBack userCallBack) {
-        progressDialog.show();
-        new StoreUserDataAsyncTask(user, userCallBack).execute();
-
-    }
-
-    public void fetchUserDataInBackground(User user, GetUserCallBack callBack) {
-        progressDialog.show();
-        new fetchUserDataAsyncTask(user, callBack).execute();
-
-    }
-
-    public class StoreUserDataAsyncTask extends AsyncTask<Void, Void, Void> {
-
-        User user;
-        GetUserCallBack userCallBack;
-
-        public StoreUserDataAsyncTask(User user, GetUserCallBack userCallBack) {
-            this.user = user;
-            this.userCallBack = userCallBack;
-
-        }
-
-        @Override
-        protected Void doInBackground(Void... params) {
-
-            ArrayList<BasicNameValuePair> dataToSend = new ArrayList<>();
-            dataToSend.add(new BasicNameValuePair("name", user.name));
-            dataToSend.add(new BasicNameValuePair("age", user.age + ""));
-            dataToSend.add(new BasicNameValuePair("email", user.email));
-            dataToSend.add(new BasicNameValuePair("password", user.password));
-
-            HttpParams httpRequestParams = new BasicHttpParams();
-            HttpConnectionParams.setConnectionTimeout(httpRequestParams, CONNECTION_TIMEOUT);
-            HttpConnectionParams.setSoTimeout(httpRequestParams, CONNECTION_TIMEOUT);
-
-            HttpClient client = new DefaultHttpClient();
-            HttpPost post = new HttpPost(SERVER_ADDRESS + "Register.php");
-
-            try {
-                post.setEntity(new UrlEncodedFormEntity(dataToSend));
-                client.execute(post);
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-
-
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(Void aVoid) {
-            progressDialog.dismiss();
-            userCallBack.done(null);
-            super.onPostExecute(aVoid);
-        }
-    }
-
-    public class fetchUserDataAsyncTask extends AsyncTask<Void, Void, User> {
-
-        User user;
-        GetUserCallBack userCallBack;
-
-        public fetchUserDataAsyncTask(User user, GetUserCallBack userCallBack) {
-            this.user = user;
-            this.userCallBack = userCallBack;
-
-        }
-
-        @Override
-        protected User doInBackground(Void... params) {
-
-            ArrayList<BasicNameValuePair> dataToSend = new ArrayList<>();
-            dataToSend.add(new BasicNameValuePair("email", user.email));
-            dataToSend.add(new BasicNameValuePair("password", user.password));
-
-            HttpParams httpRequestParams = new BasicHttpParams();
-            HttpConnectionParams.setConnectionTimeout(httpRequestParams, CONNECTION_TIMEOUT);
-            HttpConnectionParams.setSoTimeout(httpRequestParams, CONNECTION_TIMEOUT);
-
-            HttpClient client = new DefaultHttpClient();
-            HttpPost post = new HttpPost(SERVER_ADDRESS + "FetchUserData.php");
-
-            User returnedUser = null;
-            try {
-                post.setEntity(new UrlEncodedFormEntity(dataToSend));
-                HttpResponse httpResponse = client.execute(post);
-
-                HttpEntity entity = httpResponse.getEntity();
-                String result = EntityUtils.toString(entity);
-                JSONObject jsonObject = new JSONObject(result);
-
-                if(jsonObject.length() == 0){
-                    returnedUser = null;
-                }else {
-                    String name = jsonObject.getString("name");
-                    int age = jsonObject.getInt("age");
-
-                    returnedUser = new User(name, age, user.email, user.password);
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-
-            return null;
-        }
-        @Override
-        protected void onPostExecute(User returnedUser) {
-            progressDialog.dismiss();
-            userCallBack.done(returnedUser);
-            super.onPostExecute(returnedUser);
-        }
-    }
-}
