+++ app/src/main/java/com/papfree/bloodlife/UI/MainActivity.java
@@ -1,10 +1,13 @@
-package com.papfree.bloodlife;
+package com.papfree.bloodlife.UI;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import com.papfree.bloodlife.R;
+import com.papfree.bloodlife.UI.Login.LoginActivity;
 
 
 public class MainActivity extends ActionBarActivity {
