+++ app/src/main/java/com/papfree/bloodlife/Utils/Constants.java
@@ -1,0 +1,6 @@
+package com.papfree.bloodlife.Utils;
+
+public final class Constants {
+
+    public static final String FIREBASE_URL = "https://bloodlife.firebaseio.com";
+}
