+++ app/src/main/java/com/papfree/bloodlife/Utils/Utils.java
@@ -1,0 +1,24 @@
+package com.papfree.bloodlife.Utils;
+
+import android.content.Context;
+
+import java.text.SimpleDateFormat;
+
+
+public class Utils {
+
+    /**
+     * Format the timestamp with SimpleDateFormat
+     */
+    public static final SimpleDateFormat SIMPLE_DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd HH:mm");
+    private Context mContext = null;
+
+
+    /**
+     * Public constructor that takes mContext for later use
+     */
+    public Utils(Context con) {
+        mContext = con;
+    }
+
+}
