+++ app/src/main/res/layout/activity_signup.xml
@@ -101,7 +101,7 @@
             android:layout_marginTop="24dp"
             android:layout_marginBottom="24dp"
             android:padding="12dp"
-            android:text="Create Account"/>
+            android:text="@string/title_activity_create_account"/>
 
         <TextView android:id="@+id/link_login"
             android:layout_width="fill_parent"
