+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,5 @@
 
     <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
+    <string name="title_activity_create_account">Create Account</string>
 </resources>
